Use base class for tests
