+++ src/AutoMapper.Collection.EntityFrameworkCore.Tests/EntityFramworkCoreUsingCtorTests.cs
@@ -8,7 +8,7 @@ using Xunit;
 
 namespace AutoMapper.Collection.EntityFrameworkCore.Tests
 {
-    public class EntityFramworkCoreUsingCtorTests
+    public class EntityFramworkCoreUsingCtorTests : EntityFramworkCoreTestsBase
     {
         public EntityFramworkCoreUsingCtorTests()
         {
@@ -21,71 +21,18 @@ namespace AutoMapper.Collection.EntityFrameworkCore.Tests
             });
         }
 
-        [Fact]
-        public void Should_Persist_To_Update()
+        protected override DBContextBase GetDbContext()
         {
-            var db = new DB();
-            db.Things.Add(new Thing { Title = "Test2" });
-            db.Things.Add(new Thing { Title = "Test3" });
-            db.Things.Add(new Thing { Title = "Test4" });
-            db.SaveChanges();
-
-            Assert.Equal(3, db.Things.Count());
-
-            var item = db.Things.First();
-
-            db.Things.Persist().InsertOrUpdate(new ThingDto { ID = item.ID, Title = "Test" });
-            Assert.Equal(1, db.ChangeTracker.Entries<Thing>().Count(x => x.State == EntityState.Modified));
-
-            Assert.Equal(3, db.Things.Count());
-
-            db.Things.First(x => x.ID == item.ID).Title.Should().Be("Test");
+            return new DB();
         }
 
-        [Fact]
-        public void Should_Persist_To_Insert()
-        {
-            var db = new DB();
-            db.Things.Add(new Thing { Title = "Test2" });
-            db.Things.Add(new Thing { Title = "Test3" });
-            db.Things.Add(new Thing { Title = "Test4" });
-            db.SaveChanges();
-
-            Assert.Equal(3, db.Things.Count());
-
-            db.Things.Persist().InsertOrUpdate(new ThingDto { Title = "Test" });
-            Assert.Equal(3, db.Things.Count());
-            Assert.Equal(1, db.ChangeTracker.Entries<Thing>().Count(x => x.State == EntityState.Added));
-
-            db.SaveChanges();
-
-            Assert.Equal(4, db.Things.Count());
-
-            db.Things.OrderByDescending(x => x.ID).First().Title.Should().Be("Test");
-        }
-
-        public class DB : DbContext
+        public class DB : DBContextBase
         {
             protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
             {
                 optionsBuilder.UseInMemoryDatabase("EfTestDatabase" + Guid.NewGuid());
                 base.OnConfiguring(optionsBuilder);
             }
-
-            public DbSet<Thing> Things { get; set; }
-        }
-
-        public class Thing
-        {
-            public int ID { get; set; }
-            public string Title { get; set; }
-            public override string ToString() { return Title; }
-        }
-
-        public class ThingDto
-        {
-            public int ID { get; set; }
-            public string Title { get; set; }
         }
     }
 }
